edited avrgirl serial connection, added hover for save, edited go back warming
