+++ client/src/components/DayPanels/DayPanels.less
@@ -20,7 +20,7 @@ p {
 }
 
 .compilePop {
-  margin-top: 11px !important;
+  margin-top: 12px !important;
 }
 
 .popup {
@@ -49,16 +49,19 @@ p {
 }
 
 .ModalCompile {
-  top: 80%;
+  top: 85%;
   right: 3vh;
 }
 .ModalCompile2 {
-  top: 80%;
+  top: 85%;
   right: 12vh;
 }
 .ModalCompile3 {
-  top: 80%;
+  top: 85%;
   right: 5vh;
+}
+.ModalCompile4 {
+  top: 100%;
 }
 
 #action-btn-container {
